requirements.txt
ccxt>=4.0.0
pandas>=2.0.0
numpy>=1.24.0
python-telegram-bot>=20.0
python-dotenv>=1.0.0
pybit>=5.0.0
google-generativeai>=0.8.0
matplotlib>=3.8.0
mplfinance>=0.12.10b0
